メモリリークを起こす重大問題を修正

+++ src/matrix.h
@@ -2,12 +2,12 @@
 #define MATRIX_H
 
 typedef struct { 
-    int rows; 
-    int cols;
+    size_t rows; 
+    size_t cols;
     double *data;
  } Matrix;
 
-Matrix *create_matrix(int rows, int cols);
+Matrix *create_matrix(size_t rows, size_t cols);
 int print_matrix(Matrix *data);
 Matrix *add_matrix(Matrix *data_x, Matrix *data_y);
 void free_matrix(Matrix *data);
